requirements.txt
streamlit
openai
python-dotenv
stripe
bcrypt
requests
matplotlib
pandas
pydub
duckduckgo-search
google-generativeai
Pillow
fpdf
reportlab
beautifulsoup4
youtube-transcript-api